Use g_object_new to chain up factory object construction
Previously, the ->constructor on the parent class was used and
the GType was simply replaced by the desired subtype. However,
this then skips proper construction of the subtype (i.e. no
constructor and no constructed). This will solve that problem by
explicitly calling g_object_newv again for the subtype.