Implement ggit_diff_line_get_text
Retrieve the encoding of a file from either the gui.encoding config
value, or (with higher priority) the encoding file attribute. If set,
ggit_diff_line_get_text will convert from that encoding to utf-8 and
ensures a utf-8 encoded string is returned. Characters that cannot be
decoded will contain the fallback ? character.