meson: Rename the package includedir
The `includedir` used for the headers of the package itself has been
usually named `pkgincludedir`.

This patch creates a new variable to hold the name along the api
version and uses this to create a new `pkgincludedir` variable
which is going to be used instead of the `real includedir`.

The subdir parameter of install headers has also been changed to
use the path relative to the system header directory instead of
the absolute path.

https://bugzilla.gnome.org/show_bug.cgi?id=792699