meson: Simplified introspection option check
The build of the introspection data depends on the existence of the
GObject library, the `g-ir-scanner` program, and the build not being
a cross build.

The check to the GObject library has been removed, because it is
required for compiling libgit2. The gir variable has also been
removed because there is no need to store the intermediate state.

Finally, the `build_gir` variable has been renamed to remain
consistent with other similar variables like `enable_ssh`.

https://bugzilla.gnome.org/show_bug.cgi?id=792699