meson: Improve gtkdoc documentation generation
The gtkdoc option has been changed to follow the meson porting
guidelines[0].

A greater use of different variables is done to avoid any typo when
rewriting the different strings that are needed for `gtkdoc`
function parameters.

When querying the glib dependency prefix, now the variable holding
the previous dependency search is reused.

The list of private headers has also been removed, and now the
list present in meson build from the `libgit2-glib` directory is
reused.

Finally the links present in the generated documentation are fixed
by using the `--html-dir` option.

https://bugzilla.gnome.org/show_bug.cgi?id=792699

[0] https://wiki.gnome.org/Initiatives/GnomeGoals/MesonPorting