meson: Reorganize included directories
The `libgit2_glib_inc` variable holds two directories, the source
root and also the `libgit2-glib` directory. However, this is not
necessary, because a variable for the source root already exists,
which is called `core_inc`, and the example programs which use
the `libgit2_glib_inc` variable don't need the source root include.

Now the `libgit2_glib_inc` variable only holds the `libgit2-glib`
directory.

https://bugzilla.gnome.org/show_bug.cgi?id=792699