meson: Remove unnecessary introspection options
Since meson 0.43, the `generate_gir` function has a parameter
called `header` which can be used as the name of the header to be
included by the library.

This patch uses this parameter to avoid using the `--include-header`
separately.

The `--identifier-prefix` has also been removed because the
`identifier_prefix` from `generate_gir` is already used.

https://bugzilla.gnome.org/show_bug.cgi?id=792699