meson: Fix DESTDIR use in Windows
The use of DESTDIR in Windows produces some errors[0]. To fix this
issue a different approach has been used for paths. Instead of using
absolute paths, relative paths are used as meson does. One of this
relative paths is used along with MESON_INSTALL_DESTDIR_PREFIX to
avoid any issues.

[0] https://github.com/Alexpux/MINGW-packages/pull/4560#issuecomment-431294899