meson: Improve documentation generation
GtkDoc generated documentation takes advantage of different new
meson's improvements.

It uses the `gtkdoc_html_dir` function to set the paths to be fixed.
It also makes use of the `ignore_headers` parameter. Finally, it
uses the `version.xml` file target as a content file instad of
assuming the same build location.