meson_vapi_link.py: Improve calculation of vapidir
The script gets the *absolute* datadir passed as its argument. This
means os.path.join will throw away the MESON_INSTALL_DESTDIR_PREFIX,
breaking the symlinking when DESTDIR is set.

Also handle the case where the datadir is not inside the prefix by
prepending the destdir ourselves. We make the datadir relative to the
root by stripping the drive letter (if any) and the initial separator.