meson: Remove workaround comment and add proper fix to NDEBUG
Due to an issue with meson's `buildtype` and `b_ndebug` options[0],
the `b_ndebug` option was converted to a combo option.

This forced `libgit2-glib` to use a workaround that has ended being
the standard meson's behavior. The `b_ndebug` option has also gained
another option called `if-release`.

Due to this changes, the workaround comment has been removed and
proper fix has been used in place.

[0] https://github.com/mesonbuild/meson/pull/1896