Revert "Use python3"
This reverts commit 8a1da431dcbf4eda612b5189201f502fbf122726.

The commit removed some use of the `python` module to be able to
explicitly compile with `python3` in the post install script
because although meson was using `python` module properly the post
install script was calling `python` binary that might not point to
`python3` binary.

The commit has been reverted and proper solution has been applied by
passing the `python3` installation path used by meson to the post
install script.

The `python3-gobject-devel` has also been added as a requirement for
the CI step.